+++ ReleaseNotes/release_notes.md
@@ -3,6 +3,7 @@
   1. add `+=` and `-=` operators for `Ostap::Math::HermiteSum`
   1. re-introduce `ostapfitting.funbasic.func_factory` for better backward compatibility 
   1. add forward declarations for Karlin-Shapley and Karlin-Studden 
+  1. add more constructors to `Ostap::Math::Polynomial`
 
 ## Backward incompatible: